requirements.txt
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
dateparser==0.7.6
gevent==20.6.2
greenlet==0.4.16
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.10
websocket==0.2.1
zope.event==4.4
zope.interface==5.1.0
